+++ sdk/monitor/azure-monitor-query/CHANGELOG.md
@@ -23,6 +23,7 @@
 - `LogsBatchQuery` doesn't accept a `request_id` anymore.
 - `MetricsMetadataValues` is removed. A dictionary is used instead.
 - `time_stamp` is renamed to `timestamp` in `MetricValue` type.
+- `AggregationType` is renamed to `MetricAggregationType`.
 
 ### Bugs Fixed